+++ sample/platform/STM32/OnBoardSDK_STM32/Linker/hal/osdkhal_stm32.h
@@ -49,10 +49,10 @@ extern "C" {
 
 /* Exported functions --------------------------------------------------------*/
 
-E_OsdkStat OsdkSTM32_UartSendData(const T_HalObj *obj, const uint8_t *pBuf, uint16_t bufLen);
-E_OsdkStat OsdkSTM32_UartReadData(const T_HalObj *obj, uint8_t *pBuf, uint16_t *bufLen);
-E_OsdkStat OsdkSTM32_UdpSendData(const T_HalObj *obj, const uint8_t *pBuf, uint16_t bufLen);
-E_OsdkStat OsdkSTM32_UdpReadData(const T_HalObj *obj, uint8_t *pBuf, uint16_t *bufLen);
+E_OsdkStat OsdkSTM32_UartSendData(const T_HalObj *obj, const uint8_t *pBuf, uint32_t bufLen);
+E_OsdkStat OsdkSTM32_UartReadData(const T_HalObj *obj, uint8_t *pBuf, uint32_t *bufLen);
+E_OsdkStat OsdkSTM32_UdpSendData(const T_HalObj *obj, const uint8_t *pBuf, uint32_t bufLen);
+E_OsdkStat OsdkSTM32_UdpReadData(const T_HalObj *obj, uint8_t *pBuf, uint32_t *bufLen);
 E_OsdkStat OsdkSTM32_UartInit(const char *port, const int baudrate, T_HalObj *obj);
 E_OsdkStat OsdkSTM32_UdptInit(const char *addr, uint16_t port, T_HalObj *obj);
